refactor: move WordStatus to entities/word/ui
