+++ src/2_widgets/WordMainInfo/WordInfo/WordOriginal/WordOriginal.tsx
@@ -7,7 +7,7 @@ import {
 	ButtonIcon,
 } from '@ui/index';
 import { IWord } from '../WordInfo';
-import { WordStatus } from '@features/getWordStatus/ui';
+import { WordStatus } from '@/4_entities/word/ui';
 import SvgLang from '@assets/icons/wordProfile/language.svg?react';
 import SvgComment from '@assets/icons/wordProfile/comment.svg?react';
 import SvgShare from '@assets/icons/wordProfile/arrow.svg?react';
