+++ src/4_entities/word/ui/WordStatus/WordStatus.tsx
