+++ src/4_entities/word/ui/WordStatus/index.ts
