+++ src/4_entities/word/ui/index.ts
